requirements.txt

python-dotenv
requests
lumaai
openai
numpy
pandas
matplotlib
flask
cryptography

